Revert "Extract formula into own file. Add DATE_VALUE to formulas. Replace FLIP"
This reverts commit 59b2e1d7ca45b907b5568b82cffca803fd0a6f23.

Change-type: patch
