+++ lib/index.js
@@ -5,20 +5,35 @@
  */
 
 const _ = require('lodash')
+const formula = require('@formulajs/formulajs')
 const staticEval = require('static-eval')
 const esprima = require('esprima')
 const assert = require('@balena/jellyfish-assert')
 const card = require('./card')
-const formula = require('./formula')
-const {
-	checkForSpecialCases
-} = require('./special-cases')
 
 /**
  * Jellyscript module.
  *
  * @module jellyscript
  */
+
+formula.PARTIAL = _.partial
+formula.FLIP = _.flip
+formula.PROPERTY = _.get
+
+formula.REGEX_MATCH = (regex, string) => {
+	return string.match(regex)
+}
+
+formula.AGGREGATE = (list, func) => {
+	return Array.from(list.reduce((accumulator, element) => {
+		for (const value of func(element)) {
+			accumulator.add(value)
+		}
+
+		return accumulator
+	}, new Set()))
+}
 
 const runAST = (ast, options) => {
 	return staticEval(ast, Object.assign({
@@ -133,10 +148,13 @@ exports.getTypeTriggers = (typeCard) => {
 	for (const path of card.getFormulasPaths(typeCard.data.schema)) {
 		const ast = esprima.parse(path.formula).body[0].expression
 
-		const specialCase = checkForSpecialCases(ast)
-		if (specialCase) {
-			const literal = ast.arguments[1] || ast.arguments[0].arguments[1]
-			const arg = runAST(literal, {
+		// Aggregating over $events is a special case
+		if (ast.type === 'CallExpression' &&
+				ast.callee.type === 'Identifier' &&
+				ast.callee.name === 'AGGREGATE' &&
+				ast.arguments[0].type === 'Identifier' &&
+				ast.arguments[0].name === '$events') {
+			const arg = runAST(ast.arguments[1], {
 				context: {},
 				input: {}
 			})
