+++ lib/index.spec.js
@@ -1,0 +1,361 @@
+/*
+ * Copyright (C) Balena.io - All Rights Reserved
+ * Unauthorized copying of this file, via any medium is strictly prohibited.
+ * Proprietary and confidential.
+ */
+
+const ava = require('ava')
+const jellyscript = require('./index')
+
+ava('.evaluate(): should return null if no input', (test) => {
+	const result = jellyscript.evaluate('POW(input, 2)', {
+		context: {},
+		input: null
+	})
+
+	test.deepEqual(result, {
+		value: null
+	})
+})
+
+ava('.evaluate(): should resolve a number formula', (test) => {
+	const result = jellyscript.evaluate('POW(input, 2)', {
+		context: {
+			number: 2
+		},
+		input: 2
+	})
+
+	test.deepEqual(result, {
+		value: 4
+	})
+})
+
+ava('.evaluate(): should resolve composite formulas', (test) => {
+	const result = jellyscript.evaluate('MAX(POW(input, 2), POW(input, 3))', {
+		context: {
+			number: 2
+		},
+		input: 2
+	})
+
+	test.deepEqual(result, {
+		value: 8
+	})
+})
+
+ava('.evaluate(): should access other properties from the card', (test) => {
+	const result = jellyscript.evaluate('ADD(this.value1, this.value2)', {
+		context: {
+			value1: 2,
+			value2: 3
+		},
+		input: 0
+	})
+
+	test.deepEqual(result, {
+		value: 5
+	})
+})
+
+ava('.evaluate(): should evaluate a function', (test) => {
+	const result = jellyscript.evaluate('FLIP(POW)', {
+		context: {},
+		input: 0
+	})
+
+	test.is(result.value(2, 2), 4)
+	test.is(result.value(3, 2), 8)
+})
+
+ava('AGGREGATE: should ignore duplicates', (test) => {
+	const result = jellyscript.evaluate('AGGREGATE(input, PARTIAL(FLIP(PROPERTY), "mentions"))', {
+		context: {},
+		input: [
+			{
+				mentions: [ 'foo', 'bar' ]
+			},
+			{
+				mentions: [ 'bar', 'baz' ]
+			},
+			{
+				mentions: [ 'baz', 'qux' ]
+			}
+		]
+	})
+
+	test.deepEqual(result, {
+		value: [ 'foo', 'bar', 'baz', 'qux' ]
+	})
+})
+
+ava('AGGREGATE: should aggregate a set of object properties', (test) => {
+	const result = jellyscript.evaluate('AGGREGATE(input, PARTIAL(FLIP(PROPERTY), "mentions"))', {
+		context: {},
+		input: [
+			{
+				mentions: [ 'foo' ]
+			},
+			{
+				mentions: [ 'bar' ]
+			}
+		]
+	})
+
+	test.deepEqual(result, {
+		value: [ 'foo', 'bar' ]
+	})
+})
+
+ava('REGEX_MATCH: should extract a set of mentions', (test) => {
+	const result = jellyscript.evaluate('REGEX_MATCH(/(@[a-zA-Z0-9-]+)/g, input)', {
+		context: {},
+		input: 'Hello @johndoe, and @janedoe'
+	})
+
+	test.deepEqual(result, {
+		value: [ '@johndoe', '@janedoe' ]
+	})
+})
+
+ava('REGEX_MATCH: should consider duplicates', (test) => {
+	const result = jellyscript.evaluate('REGEX_MATCH(/(@[a-zA-Z0-9-]+)/g, input)', {
+		context: {},
+		input: 'Hello @johndoe, and @janedoe, and @johndoe'
+	})
+
+	test.deepEqual(result, {
+		value: [ '@johndoe', '@janedoe', '@johndoe' ]
+	})
+})
+
+ava('.evaluateObject() should evaluate a number formula', async (test) => {
+	const result = jellyscript.evaluateObject({
+		type: 'object',
+		properties: {
+			foo: {
+				type: 'number',
+				$$formula: 'POW(input, 2)'
+			}
+		}
+	}, {
+		foo: 3
+	})
+
+	test.deepEqual(result, {
+		foo: 9
+	})
+})
+
+ava('.evaluateObject() should evaluate a formula in a $ prefixed property', async (test) => {
+	const result = jellyscript.evaluateObject({
+		type: 'object',
+		properties: {
+			$foo: {
+				type: 'number',
+				$$formula: 'POW(input, 2)'
+			}
+		}
+	}, {
+		$foo: 3
+	})
+
+	test.deepEqual(result, {
+		$foo: 9
+	})
+})
+
+ava('.evaluateObject() should evaluate a formula in a $$ prefixed property', async (test) => {
+	const result = jellyscript.evaluateObject({
+		type: 'object',
+		properties: {
+			$$foo: {
+				type: 'number',
+				$$formula: 'POW(input, 2)'
+			}
+		}
+	}, {
+		$$foo: 3
+	})
+
+	test.deepEqual(result, {
+		$$foo: 9
+	})
+})
+
+ava('.evaluateObject() should ignore missing formulas', async (test) => {
+	const result = jellyscript.evaluateObject({
+		type: 'object',
+		properties: {
+			foo: {
+				type: 'number',
+				$$formula: 'POW(input, 2)'
+			}
+		}
+	}, {
+		bar: 3
+	})
+
+	test.deepEqual(result, {
+		bar: 3
+	})
+})
+
+ava('.evaluateObject() should not ignore the zero number as missing', async (test) => {
+	const result = jellyscript.evaluateObject({
+		type: 'object',
+		properties: {
+			foo: {
+				type: 'number',
+				$$formula: 'MAX(input, 2)'
+			}
+		}
+	}, {
+		foo: 0
+	})
+
+	test.deepEqual(result, {
+		foo: 2
+	})
+})
+
+ava('.evaluateObject() should evaluate nested formulas', async (test) => {
+	const result = jellyscript.evaluateObject({
+		type: 'object',
+		properties: {
+			foo: {
+				type: 'object',
+				properties: {
+					bar: {
+						type: 'object',
+						properties: {
+							baz: {
+								type: 'number',
+								$$formula: 'POW(input, 2)'
+							}
+						}
+					}
+				}
+			}
+		}
+	}, {
+		foo: {
+			bar: {
+				baz: 2
+			}
+		}
+	})
+
+	test.deepEqual(result, {
+		foo: {
+			bar: {
+				baz: 4
+			}
+		}
+	})
+})
+
+ava('.evaluateObject() should not do anything if the schema has no formulas', async (test) => {
+	const result = jellyscript.evaluateObject({
+		type: 'object',
+		properties: {
+			foo: {
+				type: 'string'
+			},
+			bar: {
+				type: 'number'
+			}
+		}
+	}, {
+		foo: '1',
+		bar: 2
+	})
+
+	test.deepEqual(result, {
+		foo: '1',
+		bar: 2
+	})
+})
+
+ava('.getTypeTriggers() should report back watchers when aggregating events', async (test) => {
+	const triggers = jellyscript.getTypeTriggers({
+		slug: 'thread',
+		type: 'type@1.0.0',
+		version: '1.0.0',
+		data: {
+			schema: {
+				type: 'object',
+				properties: {
+					data: {
+						type: 'object',
+						properties: {
+							mentions: {
+								type: 'array',
+								$$formula: 'AGGREGATE($events, "data.mentions")'
+							}
+						}
+					}
+				}
+			}
+		}
+	})
+
+	test.deepEqual(triggers, [
+		{
+			type: 'triggered-action@1.0.0',
+			version: '1.0.0',
+			slug: 'triggered-action-thread-data-mentions',
+			requires: [],
+			capabilities: [],
+			active: true,
+			tags: [],
+			links: {},
+			markers: [],
+			data: {
+				async: true,
+				type: 'thread@1.0.0',
+				action: 'action-set-add@1.0.0',
+				target: {
+					$eval: 'source.links[\'is attached to\'][0].id'
+				},
+				arguments: {
+					property: 'data.mentions',
+					value: {
+						$if: 'source.data.mentions',
+						then: {
+							$eval: 'source.data.mentions'
+						},
+						else: []
+					}
+				},
+				filter: {
+					type: 'object',
+					$$links: {
+						'is attached to': {
+							type: 'object',
+							required: [ 'type' ],
+							properties: {
+								type: {
+									type: 'string',
+									const: 'thread@1.0.0'
+								}
+							}
+						}
+					},
+					required: [ 'data' ],
+					properties: {
+						data: {
+							type: 'object',
+							required: [ 'payload' ],
+							properties: {
+								payload: {
+									type: 'object'
+								}
+							}
+						}
+					}
+				}
+			}
+		}
+	])
+})
